feat(heursitics): renamed files for efficiency
